Sixth Commit)
I completely completed the balance in the store, also made a bonus panel and it already has an interface, added a couple of sounds, wrote a couple of scripts, completed the starting scene.)

+++ minecraftClicker/Assets/Scripts/SaveGame/Save.cs
@@ -1,9 +1,9 @@
 [System.Serializable]
 public class Save 
 {
-    public double Score;
+    public int Score;
     public int[] ConstBonus;
-    public double ClickScore;
+    public int ClickScore;
     public bool IsAchievement;
     public int AchievementMax;
 }
